Nova versão, agora com lista de clientes aguardando envio.

+++ client_window.py
@@ -1,6 +1,6 @@
 import sys
 from PyQt5.QtWidgets import QApplication, QWidget, QVBoxLayout, QHBoxLayout, QTableWidget, QTableWidgetItem, \
-    QPushButton, QDialog, QFormLayout, QLineEdit, QCheckBox, QComboBox, QLabel
+    QPushButton, QDialog, QFormLayout, QLineEdit, QCheckBox, QComboBox, QLabel, QMessageBox
 from PyQt5.QtCore import Qt, QDate
 from lojaDB import buscar_compras, registrar_compra, editar_compra, excluir_compra, buscar_cliente_por_id
 from utils import criar_seletor_data
@@ -11,7 +11,7 @@ class ClientWindow(QWidget):
         self.client_id = client_id
         self.client_name = self.get_client_name()
         self.setWindowTitle('Detalhes do Cliente')
-        self.setGeometry(100, 100, 900, 600)
+        self.setGeometry(100, 100, 1180, 600)
         self.initUI()
 
     def initUI(self):
@@ -57,30 +57,7 @@ class ClientWindow(QWidget):
         self.load_compras()
         self.table_compras.resizeColumnsToContents()  # Ajustar após carregar os dados
 
-        # Aplicar estilo diretamente
-        self.setStyleSheet("""
-                            QWidget {
-                                background-color: #f0f0f0;
-                            }
 
-                            QHeaderView::section {
-                                background-color: #f0f0f0; 
-                                color: black; 
-                            }
-                            QPushButton {
-                                background-color: #3c4c7d;
-                                color: white;
-                                border-radius: 5px;
-                                padding: 10px;
-                                font-size: 16px;
-                            }
-                            QPushButton:hover {
-                                background-color: #24346c;
-                            }
-                            QPushButton:pressed {
-                                background-color: #465184;
-                            }
-                        """)
 
     def get_client_name(self):
         client_data = buscar_cliente_por_id(self.client_id)
@@ -108,7 +85,10 @@ class ClientWindow(QWidget):
     def adicionar_compra(self):
         dialog = CompraDialog(self, self.client_id)
         if dialog.exec_():
+            print("salvando a nova compra...")
+            dialog.save_compra()
             self.load_compras()
+            print("Nova compra adicionada com sucesso.")
 
     def editar_compra(self):
         selected_row = self.table_compras.currentRow()
@@ -152,12 +132,27 @@ class ClientWindow(QWidget):
         selected_row = self.table_compras.currentRow()
         if selected_row != -1:
             compra_id = self.table_compras.item(selected_row, 0).text()  # Ajustar para pegar o ID correto
-            try:
-                excluir_compra(compra_id)
-                print(f"Compra com ID {compra_id} excluída com sucesso.")
-                self.load_compras()
-            except Exception as e:
-                print(f"Erro ao excluir compra: {e}")
+            #Caixa de confirmação
+            msg_box = QMessageBox()
+            msg_box.setIcon(QMessageBox.Warning)
+            msg_box.setText("Tem certeza que deseja excluir esta compra?")
+            msg_box.setInformativeText(f"A compra com ID {compra_id} será permanentemente apagada")
+            msg_box.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
+            msg_box.setDefaultButton(QMessageBox.No)
+
+            #Exibir mensagem e capturar a resposta
+            response = msg_box.exec_()
+            if response == QMessageBox.Yes:
+                try:
+                    excluir_compra(compra_id)
+                    print(f"Compra com ID {compra_id} excluída com sucesso.")
+                    self.load_compras()
+                except Exception as e:
+                    print(f"Erro ao excluir compra: {e}")
+            else:
+                print(f"Exclusão da compra cancelada.")
+        else:
+            print("Nenhuma compra selecionada para exclusão")
 
 
 class CompraDialog(QDialog):
@@ -228,6 +223,32 @@ class CompraDialog(QDialog):
             'codigo_rastreio': self.input_codigo_rastreio.text(),
             'enviado': self.checkbox_enviado.isChecked()
         }
+
+    def save_compra(self):
+        try:
+            dados_compra = self.get_dados_compra()
+            print(f"salvando compra: {dados_compra}")  # Log para ver os dados obtidos
+            if self.compra_data:
+                editar_compra(
+                    self.compra_data[1],  # id da compra
+                    dados_compra['data_venda'],
+                    dados_compra['produto'],
+                    dados_compra['valor_venda'],
+                    dados_compra['modo_pagamento'],
+                    dados_compra['data_pagamento'],
+                    dados_compra.get('data_envio', None),
+                    dados_compra.get('codigo_rastreio', None),
+                    1 if dados_compra.get('enviado', False) else 0
+                )
+                print("Compra editada com sucesso.")  # Verifique se chega aqui ao editar
+            else:
+                print("Tentando registrar nova compra...")  # Verifique se o fluxo chega aqui
+                registrar_compra(self.client_id, **dados_compra)
+
+            self.accept()
+        except Exception as e:
+            print(f"Erro ao salvar compra: {e}")
+
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
